requirements.txt
django == 4.2
pillow
django-admin-interface
django-bootstrap5
djangorestframework
django-cors-headers